refac: use interface instead of implementation
