+++ src/main/java/nextstep/auth/application/AuthService.java
@@ -1,9 +1,8 @@
 package nextstep.auth.application;
 
 import nextstep.auth.application.dto.GithubProfileResponse;
-import nextstep.member.application.MemberService;
-import nextstep.member.application.dto.MemberRequest;
 import nextstep.auth.application.dto.TokenResponse;
+import nextstep.auth.domain.UserDetail;
 import org.springframework.stereotype.Service;
 
 @Service
@@ -28,8 +27,8 @@ public class AuthService {
         final String accessToken = githubClient.requestGithubToken(code);
         final GithubProfileResponse profile = githubClient.requestGithubProfile(accessToken);
 
-        userDetailService.findOrCreate(profile.getEmail());
+        final UserDetail userDetail = userDetailService.findOrCreate(profile.getEmail());
 
-        return new TokenResponse(tokenService.createToken(profile.getEmail(), code).getAccessToken());
+        return new TokenResponse(tokenService.createToken(userDetail.getEmail(), userDetail.getPassword()).getAccessToken());
     }
 }
