+++ src/main/java/nextstep/auth/domain/UserDetail.java
@@ -1,15 +1,8 @@
 package nextstep.auth.domain;
 
-import nextstep.member.domain.Member;
-
 public class UserDetail {
     private String email;
     private String password;
-
-    public UserDetail(Member member) {
-        email = member.getEmail();
-        password = member.getPassword();
-    }
 
     public UserDetail(String email, String password) {
         this.email = email;
